// front/src/app/modules/shared/model/ComprasElement.ts
export interface ComprasElement {
	id: number;
    numfactura: number;
	fechafactura: string;
	proveedor: number;
	producto: number;
	catidadproductos: string;
	estado: string;
	precio: number;
	precio_unitario: number;
	usuario: number;
	num_preliminar: number;
	precio_venta: number;
	credito: number;
	numfacturap: string;
	formapago: string;
	origen: string;
	precio_venta2: number;
	precio_venta3: number;
	dinerorecibido: number;
	cambio: number;
	subtotal: number;
	ivavalor: number;
	porcentaje: number;
	bodega: number;
	registro: string;
	registro2: string;
}